Added variables property to netcdf_data
